requirements.txt
cnocr
torch
opencv-python
pillow
onnxruntime
lmdb
